feat: Implement address management and add employee credentials
This commit introduces two major features:

1.  **Employer Address Management System:**
    -   The address modal (`_address_management.blade.php`) has been updated with specific IDs for better JavaScript targeting.
    -   A new, dedicated JavaScript module (`address-handler.js`) now manages all front-end logic for the address modal, including fetching Thai address data and handling cascading dropdowns for provinces, districts, and sub-districts.
    -   The old, inline address script has been removed from the Blade partial and the new module is imported into the main `app.js`.

2.  **Employee Email and Password Fields:**
    -   A database migration has been added to include `email` and `password` columns in the `employees` table.
    -   The `Employee` model is updated to make these new fields fillable and to automatically hash passwords upon saving.
    -   The `create` and `edit` employee views now include fields for managing employee email and password.
    -   The `EmployeeController`'s `store` and `update` methods have been updated to handle the validation and persistence of the new credential fields.
