+++ app/Models/Employee.php
@@ -40,6 +40,8 @@ class Employee extends Model
         'designatedHospital',
         'startDate',
         'employeePhone',
+        'email',
+        'password',
         'employeePosition',
         'workPermitMOUGroup',
         'workPermitMOUGroupOther',
@@ -72,6 +74,7 @@ class Employee extends Model
         'employeeDob' => 'date:Y-m-d',
         'startDate' => 'date:Y-m-d',
         'terminated_at' => 'datetime',
+        'password' => 'hashed',
     ];
 
     public function employer()
