+++ resources/js/app.js
@@ -1,6 +1,7 @@
 import './bootstrap';
 import './employment-history.js';
 import './terminate-employee.js';
+import './address-handler.js';
 
 import Alpine from 'alpinejs';
